Ported zstd v1.5.7

+++ src/ZstdSharp/Unsafe/Bitstream.cs
@@ -158,7 +158,7 @@ namespace ZstdSharp.Unsafe
             BIT_flushBits(bitC);
             if (bitC->ptr >= bitC->endPtr)
                 return 0;
-            return (nuint)(bitC->ptr - bitC->startPtr + (bitC->bitPos > 0 ? 1 : 0));
+            return (nuint)(bitC->ptr - bitC->startPtr) + (nuint)(bitC->bitPos > 0 ? 1 : 0);
         }
 
         /*-********************************************************
